adaptive: move non-essentatial code to utils directory
