+++ src/core/abr/utils/pending_requests_store.ts
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-import log from "../../log";
+import log from "../../../log";
 import Manifest, {
   Adaptation,
   ISegment,
   Period,
   Representation,
-} from "../../manifest";
-import objectValues from "../../utils/object_values";
+} from "../../../manifest";
+import objectValues from "../../../utils/object_values";
 
 /**
  * Store information about pending requests, like information about:
