+++ src/core/abr/utils/representation_score_calculator.ts
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-import log from "../../log";
-import { Representation } from "../../manifest";
-import EWMA from "./utils/ewma";
+import log from "../../../log";
+import { Representation } from "../../../manifest";
+import EWMA from "./ewma";
 
 /**
  * Calculate the "maintainability score" of a given Representation:
